# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vuelos",
        "description": "Documentación de la API para gestionar vuelos.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Obtener todos los vuelos",
                "operationId": "3254b18ad6bc488e245dd7ee65b65613",
                "responses": {
                    "200": {
                        "description": "Lista de vuelos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Flight"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching flights"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Crear un vuelo",
                "operationId": "93ae933bfd08abc1b0d9b6d2a7098b7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departureCity": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "arrivalCity": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "dateDeparture": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-15T08:00:00Z"
                                    },
                                    "dateArrival": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-15T16:00:00Z"
                                    },
                                    "itinerary_id": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vuelo creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar el vuelo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error saving flight"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flights/{id}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Obtener un vuelo por ID",
                "operationId": "eadd276d0ccd85e13783b6f5b43d9ed2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vuelo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del vuelo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Flight"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vuelo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Flight not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el vuelo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching flight"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Flight": {
                "title": "Flight Schema",
                "description": "Esquema de un vuelo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "departureCity": {
                        "type": "string",
                        "example": "New York"
                    },
                    "arrivalCity": {
                        "type": "string",
                        "example": "London"
                    },
                    "dateDeparture": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T08:00:00Z"
                    },
                    "dateArrival": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T16:00:00Z"
                    },
                    "itinerary_id": {
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}